requirements.txt
pandas
numpy
rdkit-pypi
fcd